pyproject.toml
[tool.poetry]
name = "twitter"
version = "0.1.0"
description = ""
authors = ["Kelly Yuzawa <keandna@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
factory-boy = "^3.3.0"
Django = "^4.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
